requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
blinker==1.8.2
boto3==1.34.135
botocore==1.34.135
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.8
cryptography==42.0.8
fastavro==1.9.4
filelock==3.15.4
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.1
git-filter-repo==2.38.0
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
parameterized==0.9.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
types-requests==2.32.0.20240622
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
